+++ src/Automation/Automation/Properties/Resources.Designer.cs
@@ -950,5 +950,16 @@ namespace Microsoft.Azure.Commands.Automation.Properties {
                 return ResourceManager.GetString("WebhookOneTimeURL", resourceCulture);
             }
         }
+
+        /// <summary>
+        ///   Looks up a localized string similar to Insufficient user permissions. Please refer https://docs.microsoft.com/azure/role-based-access-control/built-in-roles for more details.
+        /// </summary>
+        internal static string InsufficientUserPermissions
+        {
+            get
+            {
+                return ResourceManager.GetString("InsufficientUserPermissions", resourceCulture);
+            }
+        }
     }
 }