+++ src/HealthcareApis/HealthcareApis/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 1.3.2
 * HealthcareApis cmdlets will bump up API version which may introduce breaking change. Please contact us for more information.
 
 ## Version 1.3.1